test: consolidate _propagate_hassette_logger into tests/unit/web/conftest.py
The autouse fixture was copy-pasted verbatim across five files in
tests/unit/web/. Move it to a directory-scoped conftest.py and delete the
duplicates, matching the per-subdirectory conftest convention already used
by tests/unit/{bus,core,resources,scheduler,tools,cli}/.

Drops the imports each deletion orphaned and repoints the docstring
cross-reference in tests/integration/web_api/test_auth.py, which named
tests/unit/web/test_auth.py as the canonical copy.

The integration test keeps its own definition — it sits in a separate test
tree with no shared conftest to dedupe into.

Closes #1652
