+++ tests/unit/web/conftest.py
@@ -1,0 +1,16 @@
+"""Shared fixtures for web unit tests."""
+
+import logging
+
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def _propagate_hassette_logger() -> None:
+    """Ensure the "hassette" logger propagates so caplog can see records.
+
+    Some other test in the session may have left ``propagate`` set to False (e.g. via
+    ``enable_basic_logging()``); caplog relies on propagation to the root logger. Same
+    workaround as ``tests/unit/test_validate_apps.py``.
+    """
+    logging.getLogger("hassette").propagate = True
