+++ tests/unit/web/test_auth_tokens.py
@@ -20,17 +20,6 @@ def _make_config(**overrides: Any) -> WebApiConfig:
     return WebApiConfig.model_validate(overrides)
 
 
-@pytest.fixture(autouse=True)
-def _propagate_hassette_logger() -> None:
-    """Ensure the "hassette" logger propagates so caplog can see records.
-
-    Some other test in the session may have left ``propagate`` set to False (e.g. via
-    ``enable_basic_logging()``); caplog relies on propagation to the root logger. Same
-    workaround as ``tests/unit/test_validate_apps.py``.
-    """
-    logging.getLogger("hassette").propagate = True
-
-
 class TestResolveAuthTokenExplicitConfig:
     def test_uses_configured_token_without_touching_disk(
         self, tmp_path: Path, caplog: pytest.LogCaptureFixture
